refactor: step icon story dir (#360)
