+++ src/components/Stepper/components/Step/components/StepIcon/stories/stepIcon.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import React, { Fragment } from "react";
-import { StepIcon } from "./stepIcon";
+import { StepIcon } from "../stepIcon";
 
 const meta: Meta<typeof StepIcon> = {
   component: StepIcon,
